MINOR update to refactor remark-tac to make it easier to add other plugins
