+++ packages/remark-tac/tsconfig.json
@@ -10,7 +10,6 @@
     "target": "ES2020",
     "lib": ["ES2020"],
     "module": "ES2020",
-    "moduleResolution": "node",
-    "strict": true
+    "moduleResolution": "node"
   }
 }
